Direct_decl_parser.concurrent - concurrent direc-decl parsing
Summary:
We already have `Direct_decl_parser.parse_decls`.

This diff adds `Direct_decl_parser.Concurrent`, which is designed as a convenient way for single-threaded ocaml to leverage multi-threaded Rust direct-decl-parsing.

## Single-threaded <-> multi-threaded interop

The convenience, the interop bridge between single-threaded and multi-threaded, boils down to this function which is called from the single-threaded ocaml side:
```
val enqueue_next_and_get_earlier_results :
    handle ->
    Relative_path.t list ->
    Relative_path.t * parsed_file) option
```
What is does is enqueue a list of workitems (relative-paths which we want direct-decl-parsed) onto the rust multithreaded side, and then blocks synchronously until the next available result (a parsed_file) comes from rust. The next available result might be for one of the workitems we placed just now. But it might also come from one of the earlier calls we made to this same function!

The implementation behind this is to use rayon:
1. For each relative-path, we do `rayon::spawn` to put the path (or rather, a lambda to process that path) into rayon's global workitem queue. Rayon actually stores that queue amongst threads rather than having one single storage; it uses work stealing amongst threads.
2. Prior to spawn, we increment an integer `outstanding`. This integer lives on the main thread.
3. When the spawned lambda has finished direct-decl-parsing, it sends the results over a fixed-size crossbeam channel. (Fixed-size to provide backpressure).
4. The main thread will wait for the next available result over the crossbeam channel, and decrement the `outstanding` integer. This way, if ever the integer is >0, we are guaranteed that a blocking wait on the channel will eventually be fulfilled.

## Polymorphic tag, for metadata

I actually made the function take a polymorphic tag `'a` for metadata which accompanies the workitem, and is sent back to ocaml alongside that workitem's result, without ever having been read by the rust side.
```
val enqueue_next_and_get_earlier_results :
    handle ->
    (Relative_path.t * 'a) list ->
    (Relative_path.t * 'a * parsed_file) option
```
It's solely there as a handy place to store metadata. I will use it in the next diff because (1) we know we want to prefetch the ancestors of a particular classlike which isn't itself yet in cache, (2) we'll kick off the concurrent job to fetch the decls of the file containing this classlike, (3) once we get the file's decls back, we'll want to remember which classlike we still wanted to pursue further.

## Content

I actually made the function take `content : string option`. This is to support the IDE case, where we might want to direct-decl-parse an unsaved IDE buffer.
```
type content = string option

val enqueue_next_and_get_earlier_results :
    handle ->
    (Relative_path.t * 'a * content) list ->
    (Relative_path.t * 'a * parsed_file) option
```
If it's Some then the rayon worker will direct-decl-parse the provided content; if it's None then the rayon worker will load the file off disk.

Differential Revision: D52030851

fbshipit-source-id: ccdbce00c2b32ec16363549bac55043fadd28386