Simplify get_not_subtype_fanout
Summary: Now that we have descendant deps at hand, we can simplify this function. Before the function would recursively traverse all descendants and get the NotSubtype edges from those. Now we just provide the descendant deps and iterate on those to get the NotSubtype edges ends. It's probably clearer to read the code 😅

Differential Revision: D52204374

fbshipit-source-id: 1b14f90a18b2c050baa3e8908595b403f7881dfc