Create naming_sqlite_new.rs
Summary:
**What?**
I'm going to be updating naming_sqlite.rs for a few reasons:
1. To add 2 columns for autocomplete sort text lookup for top level classes and functions
2. Remove our glean queries in autocomplete to use the naming_symbols table (should lead to a performance improvement)
3. Remove unnecessary columns for classes, funs, typedefs, modules etc from the file_info table

I'll be posting about this in Hack Engineering per nt591's rec.

**What specifically?**
I just copy naming_sqlite to here. That way I can edit this file in place without messing up saved states and rolling out a change that could be very problematic for hack users.
For review, this is just a copy that I'll be using to eventually swap out with naming_sqlite using the saved state rollout flags infra. Normally, I'd rather do a refactor/update in place but because of the risks of changing the naming tables I'm going to opt for a safer rollout which has 2 naming tables for a bit.

**What's next?**
Here are all my steps!
   1. define an entirely new saved-state-rollout-flag (Catherine's work from 2023)
   2. https://www.internalfb.com/intern/wiki/Hack_Developer_Guide/saved-state/Saved_State_Rollouts/ (**done**)
   3. there maybe exist two files "naming_sqlite.rs" and "naming_sqlite_new.rs"
      You'll still be writing to a naming.sql of the same name (**this diff!**)
   4. There's a JK saved-state-rollout-flag which determines whether we read+write from old or new schema
   5. Every few minutes, the saved-state-generator-machinery will generate (1) one saved-state for JK=false,
      (2) another saved-state for JK=true
   6. use Catherine+Bob's mass-reply tool to re-run existing sandcastle jobs on the new naming-table
   7. This would allow us to switch to lower-case column names in sqlite if you want? (e.g. old vs new) (**next diff**)

In the followups, I'll add the new columns I want and also remove the unnecessary classes columns mentioned above.

Reviewed By: nt591

Differential Revision: D52304195

fbshipit-source-id: 771d39ccf7f176655d02559b9960af1ca5c28cc3