Make SQL in naming_sqlite easier to read
Summary:
Talking to ljw1004 we both like the sql format where the keywords are capitals but the column and table names are lower case.
So in this diff I just update the casing so its easier to parse out the table names and columns and doesn't feel like a long screaming sql statement.

Reviewed By: nt591

Differential Revision: D52304978

fbshipit-source-id: ed930c9c00831a3a5122b45cdc1bbb44ca3d05ab