ensure that originalModuleName is always initialised if extShared is defined
Summary:
Methods defined in module level traits store the name of the module where they are defined in the `originalModuleName` field in the `extendedSharedData` section of the Func object.  However, it can happen that Func objects have `extShared` sections defined for other reasons (eg. if they define DocComments).  In that case the `originalModuleName` field in the `extendedSharedData` section is not initialised.

This diff ensures that whenever a Func object has a `extendedSharedData` section the `originalModuleName` is correctly initialised.  By doing so it is not necessary anymore to check if `originalModuleName` is `null` in the `moduleName()` query.

Reviewed By: jamesjwu

Differential Revision: D52094281

fbshipit-source-id: 3f569cfd08a3d4a0b32539e418be1f51098771ef