support internal methods in module level traits
Summary:
Currently module level traits cannot define internal methods or properties.  This ensures that whenever a method defined in a module level trait is invoked, it must be public and no module boundary check must be performed.

If a module level trait defines an internal method `foo`, as in:

```
/// file a.php
module A;

<<__ModuleLevelTrait>>
public trait T {
   internal function foo(): void { echo "I am foo in T\n"; }
}

/// file b.php
module B;

class C { use T; }

<<__EntryPoint>>
function bar(): void {
  $c = new C();
 $c->foo();
}
```
When executing `$c->foo()`, the runtime must fetch the module of `foo` and compare it with the module of the caller.  If module level traits can define internal methods, then fetching the module name of the callee must first check if the method representation has an `extendedData` section, if yes fetch the module name from there, otherwise fetch the module name of the unit where the method is defined.

For instance methods, the module boundary check for method invocation is performed by C++ code (and not jitted assembly) during method resolution; the relevant C++ code is already implementing the extra check, with no measured regression.

The only case where we should emit one extra condition in jitted code is in the invocation of static methods, in the lowering of `CallViolatesModuleBoundary` and `CallViolatesDeployementsBoundary`. These IR instructions can be generated only when the target is either a static method or a top-level function.   For top-level functions we should not perform the extra lookup, as a top-level function cannot be defined in a module level trait.  This diff adds the module name lookup in the check for static methods, although I could not write out a test (not even using `laundry` or reflection) that actually triggers this IR instruction in repo mode.

Reviewed By: ricklavoie, jamesjwu

Differential Revision: D51005506

fbshipit-source-id: 129c8537a4c6a5f52d8a1acd91d24caff8e2c132