use getNoCreate() in Func::load()
Summary:
Avoid creating an empty NamedFunc entry just before attempting to
autoload a function. If the autoload fails to load the given function,
we don't want the empty entry sitting around. And if autoload succeeds
but the function has a different-case name than what was requested,
we want the correct-case name to be registered.

Reviewed By: ricklavoie

Differential Revision: D52291207

fbshipit-source-id: 24acc4d4ed334466341a58b2e680121b569106ea