Consistently reserve and trim in BlobDecoder
Summary:
When deserializing maps, sets, and vecs, we want to pre-reserve the
size and trim memory usage whenever possible to save memory. We
weren't doing this consistently because not all data structures
support such operations, and the templated code had to be
generic. Instead use SFINAE to generically use the operations when
available.

Reviewed By: mdko

Differential Revision: D52349956

fbshipit-source-id: 5521176fb59c1edcc443f4fab9a484fd83d7da8d