Cache and de-dup regular expression instances for package deployment patterns
Summary:
Every unit stores a PackageInfo, and every PackageInfo contains
re2::RE2 instances corresponding to it's deployment patterns. However,
no de-duping is done, so every unit gets it's own instance. These
instances are *not* small. Moreover, the vast majority of patterns are
the same! Thus we're wasting huge amounts of memory on duplicate data.

Instead use a cache keyed by the pattern string. If we've already
created a re2::RE2 instance corresponding to the pattern, just use
that instead of creating a new one.

Differential Revision: D52350357

fbshipit-source-id: 7c6dfb4b8b197e29243e4157171a15dc550c4bde