"Unresolve" type-constraints of missing classes during init-types pass
Summary:
During flatten classes, if a type-constraint refers to a class and
that class is not known to be missing, we go ahead and resolve the
type-constraint to SubObject of that class. However, we might later
learn that that class is uninstantiable (or missing). Add logic to the
init-types pass to "unresolve" any such type-constraint back to
Unresolved. This avoids having resolved type-constraints referring to
non-existant classes.

Reviewed By: kavoor

Differential Revision: D52381474

fbshipit-source-id: 9d880e7fbf6eee4494b5e343dc13f0c1d16f401f