Fix apc_expire test
Summary: There's a race condition here - APC TTLs are based on second-resolution accuracy so if we set a TTL of 1s and happen to cross the second boundary before the fetch (unlikely * runs a lot == it happens) it will have already timed out. Since we're running 10 times with a delay of 1s on each iteration then a 3s TTL should be plenty.

Reviewed By: ricklavoie

Differential Revision: D52398509

fbshipit-source-id: b580565de0d7b3a9068a19b7f051642148ac38c1