lsp_invalidation 1/3 - strategy
Summary:
We have a problem with invalidation in the IDE - it sometimes invalidates too many things (making it slow), but sometimes not enough (making it report spurious errors, or throw an exception during typechecking which results in no IDE services for the file in question).

Last month had tried to help this with "sticky quarantine", an idea which would lead to drastically less invalidation in the common case of a user typing in a method body such that no signatures are changed hence no invalidation is needed, giving us the freedom to do more costly correct invalidation in the less-common case of signatures being changed. But the experiment failed -- the "more costly" part must have been too costly (although there wasn't enough detailed telemetry so I can't be sure).

This diff stack therefore tries to make correct invalidation cheaper. It also adds more telemetry.

## Invalidation needs

I don't know what the best solution is! ClientIdeDaemon doesn't maintain a depgraph like hh_server, and I hope it never will, so we need to figure out different strategies for efficient invalidation. I'm trying a few guesses as to what will work. I figure with telemetry data from these experiments I can figure out next steps.

If the user changed something only in a method body, sticky_quarantine already means it causes no invalidation. If they change it in a *function/const/module/typedef* signature, that's easy, because we only have to invalidate that single decl. It's only when they change something in a *class* signature that we have to worry about which folded decls to invalidate. *The correct answer is to invalidate every folded decl in cache which either is the changed class, or has the changed class as one of its ancestors.*

For class changes, that's the crux of this diff. Previously sticky_quarantine did "fast+incorrect through being too small" invalidation for keystrokes, and "correct through being vastly too large" for file-changes. This diff aims to make it "correct, and fast through being precise/small".

## Scenarios

Let's evaluate what I believe are the common cases (although this will be verified by telemetry):
1. If user types keystrokes in a method body or whitespace, the existing sticky-quarantine code detects that the file's position-free-decl-hash hasn't changed, so no invalidation is done
2. If the user types keystrokes in a class signature, then this diff lets us do fine-grained decl-diffing to detect that only this class has changed, and we'll go on to do an optimized invalidation (provided by this diff) in the case of just a single class changed.
3. If the user does Cmd+S, the existing sticky-quarantine code detects that the file's position-free-decl-hash hasn't changed, so no invalidation is done.
4. If the user runs a codegen script to generate say ten files, or does a rebase that changes hundreds of thousands of files, then we'll leave it with the heavyweight approach of simply invalidating every single folded decl.

## Fine-grained decl-diffing

This diff adds fine-grained decl-diffing to determine precisely which classes changed. There are two paths:
1. **Enter sticky quarantine** (e.g. upon keystroke). In this case, sticky quarantine already direct-decl-parsed the new decl content of the file, so we already have them in hand, so it's easy to decl-diff the new decls against the old ones (if they were present in cache).
2. **File change** (e.g. upon Cmd+S, or rebase). We have run the rust multithreaded indexer, but don't have decls in hand. In the Cmd+S case where only a single file has changed, the position-free-decl-hash of the new file content will be the same as what was already sticky quarantined, so we won't even do any invalidation. In the rebase case where maybe 10k files have changed -- including files we didn't even care about -- I don't think it would make sense to direct-decl-parse them all merely to do decl diffing, so this diff doesn't.

The rule is: *Given a list of class names which have maybe been changed, if we were given the new decl content of the changed file and found the class name in that list, and if the class name happened to be in the existing shallow-decl-cache, then we can do fine-grained decl diffing to maybe determine that the class wasn't in fact changed.*

## Optimized invalidation when just a single class is changed

Recall the goal: "if class C has changed, then we must invalidate the folded decl of C if it's in the cache, plus every single other folded decl which has C among its ancestors".

For simplicity, I decided to implement this algorithm only for a single class changed. It made the code easier to read. I don't know what the best strategy is. Here are candidates...
1. We could walk every single folded decl F in the cache, and for each one, check if C is mentioned in its ancestor-list. *Best I can think of.*
2. We could invalidate every single folded decl F in the cache. *This will be slowest*.
3. We could maintain a separate data-structure "C -> all folded decls which have C as their ancestor" but which wouldn't get evicted like the other LRU caches; instead it'd get updated whenever a folded decl gets added or removed. *I think it would be a pain to maintain this separate data structre*.
4. We could maintain  a separate data-structure "C -> immediate descendents" but again make sure this wouldn't get evicted until it was okay. *It's too confusing to think what those eviction rules would be.*

This diff uses (1).

Actually it uses (1/2) based on the JK `lsp_invalidate`, because I wanted to get an idea for how costly (2) might be. This is interesting because (1) provides the exact correct implementation, i.e. it invalidates neither too many nor too few folded decls. Therefore with (1/2) we will see telemetry for how long the invalidation takes -- hence whether it's worth investing in a faster implementation that gets the correct invalidation. And we'll see telemetry for how future typechecks perform with correct vs too-large invalidation, i.e. a "speed of light" limit on how fast future typechecks can be.

This diff creates method stubs for (1); implementation is left for later in the stack

Reviewed By: patriciamckenzie

Differential Revision: D52334870

fbshipit-source-id: b9bf9ef287b15eb59b8e90dacffca78612e86ccc