Fix stop()/join() implementation for MuxIOThreadPoolExecutor
Summary:
Currently the correct shutdown sequence is only implemented in `MuxIOThreadPoolExecutor`'s destructor, but `stop()` and `join()` should behave in the same way: call the `IOObserver`s to unregister the `EventBase`s before stopping the threads that run them. Otherwise, calling `join()` or `stop()` results in the observers being called too late.

To fix this, just move the shutdown sequence to `join()`, and delegate all the other methods to it.

Reviewed By: dmm-fb

Differential Revision: D52521695

fbshipit-source-id: b10a9aa487efae72c85ea0be9df808a7fe885424