add a global option to select the surprise flags based or the fast `fb_intercept2`
Summary: This diff proposes to add a global option to select if `fb_intercept2` is implemented with surprise flags (the current behaviour) or with instruction smashing (the "fast" behaviour).   Both implementation are kept in the codebase, so we can rollback to the surprise-flag based implementation if problems arise during the rollout of the fast `fb_intercept2`.  Also, the flag might give flexibility to measure the perf impact of the fast `fb_intercept2`.

Reviewed By: paulbiss

Differential Revision: D51847761

fbshipit-source-id: 225a29365e2bfd2e6b8cacf5d3eab075bd60fc13