fb_intercept2 via jcc/jmp smash
Summary:
Currently, whenever one method is intercepted, the Intercept bit of the surprise flags vector is set to true, and all method invocations, including those of the unintercepted methods, trigger the surprise flag handling code.

This diff revisits the implementation of `fb_intercept2` mechanism: instead of relying on surprise flags `fb_intercept2` modfies the jitted assembly code of intercepted methods, avoiding slowing down non-intercepted methods.

Strategy:

All funcentries test the surprise flags with a `cmp; jcc` instruction sequence; if any surprise flag is set, then the `jcc` invokes the surprise flag handler.  With this diff, the Intercept surprise flag (bit 53) is removed, and whenever a method is intercepted, the `jcc` instruction in the funcentry is rewritten into a `jmp` (actually, `nop; jmp` so it has the same size of `jcc` instruction).  By doing so `fb_intercept2` ensures that the surprise flag handler gets invoked only for intercepted methods.

Implementation details:

- a new instruction, `interceptjcc` is introduced; it assembles as `jcc` but whenever it is emitted in a funcentry to test the surprise flags, its TCA is stored in the global table `globalntecept`.  For this, `interceptjcc` addresses are first tracked in the `interceptTCAs` table in `CGMeta`, and then copied in the global table during relocation.  The `interceptjcc` is used when lowering `CheckSurpriseFlagsEnter` to compare the surprise flag vector;

- whenever a method is `fb_intercept2`ed, all its jitted instances in the TC have the surprise flags `jcc` instruction rewritten into `nop; jmp` (function `smashInterceptJcc` in `smashable_instrcpp`);

- if a method is already intercepted when it is jitted,  its `interceptjcc` (assembled as `jcc`) instruction is rewritten into `nop; jmp` just before its translation is published in the TC;

- the global table `globalIntercept` performs reference counting of all the requests that intercept a given function; if a previously intercepted function is not intercepted any longer by any request, then the `nop; jmp`instructions in its jitted funcentries are rewritten back into `jcc`;

- the surprise flag handler is updated so that it does not check the Intercept flag to determine if a method is intercepted, but checks rather `func->maybeIntercepted()` and, if that returns true, it validates the guess by looking for a intercept handler in the RDS.  The Intercept bit (53) is no longer used;

- similarly, the interpreter relies on the `func->maybeIntercepted()` + search for an intercept handler in the RDS check to determine if a method is intercepted or not.

The (potentially concurrent) accesses on `globalIntercept` (see `tc-intercept.cpp`):

- `startInterceptFunc`: when method `foo` is intercepted by a request, `globalIntercept[func]` counter is increased; if after the increase it is equal 1, then all the TCAs in `globalIntercept[func]` are rewritten into `nop;jmp`;
- `stopInterceptFunc`: when method `foo` is no longer intercepted by a request, `globalIntercept[func]` counter is decreased; if after the decrease it is equal 0, then all the TCAs in `globalIntercept[func]` are rewritten into `jcc`;
- `recordInterceptTCA`: whenever a funcentry for `foo` is jitted, the TCA of its `interceptjcc` is written into the TCAs of `globalIntercept[func]` during relocation; additionally, if the `globalIntercept[func]` counter is greater than 0, then the translation `interceptjcc`s are rewritten into `nop; jmp`.

The dangerous race-condition is when a method is jitted by one request, and in parallel it is intercepted by another request: we must ensure that if a method is intercept by at least one request, then all its translations have the `interceptjcc` rewritten into `nop; jmp`.  For this all the accesses to `globalIntercept` are protected by the `globalInterceptLock` and  the `globalInterceptLock` is taken by the `publishCode` function, to ensure that `interceptjcc` -> `nop; jmp` rewrite and publishing of the translation in the TC are atomic.

Reviewed By: paulbiss

Differential Revision: D48960165

fbshipit-source-id: 529ec47ce1d091a375354a976a7cdb1cabe4bf84