NamedType::getNoCreate in Class::exists()
Summary:
Calling Class::exists() with a wrong-cased name latches the wrong-case name
into the NamedType or NamedFunc table prior to attempting to autoload the symbol.

Use getNoCache() and handle null `ne` pointers to avoid latching the query name;
once a symbol is loaded we will use it's declared name as the key.

Reviewed By: aorenste

Differential Revision: D52363694

fbshipit-source-id: 96de874c3a36aad5014605cd8a36c76d2a720d94