cache FileDecls with sha1 hash as key
Summary:
Uses the SHA1 hash of the file contents as a cache key to ensure we invalidate the cache if the file changes without needing Watchman events.

To support this, I changed `FileDecls` to use a shared_ptr instead of `optional` based on the advice in D52259369. I store a shared_ptr in a process-wide cache, and increment the reference with the copy constructor on a cache hit.

Reviewed By: edwinsmith

Differential Revision: D51820433

fbshipit-source-id: 6c7dc867f1344ea868248a0d302ec96e10602b68