Respect the auto-namespace map in top-level statements
Summary:
**What**

Fix T172889179: the auto-namespace map wasn't respected in top-level statements. This is mainly relevant to Hack notebooks, where top-level statements are allowed.

**Why**

Top-level statements are used in notebooks, an should behave predictably. We use the auto-namespace map elsewhere, so should probably use it in top-level statements

**How**

**The code introduced in this diff will be deleted** when/if we switch to Rust elab which "just works" without hacks, see D52514104.

The implementation introduces a minor inconsistency in internal API shape due to an existing inconsistency:
- We pass parser options into the elaborator for top-level statements, but not for other top-level definitions

This is a consequence of an existing inconsistency:
- For most top-level definitions (classes, functions, etc.) we store a namespace environment in AST nodes.
- We do not store a namepace environment in the AST node for statements. I suspect doing so would be prohibitively expensive (and potentially confusing) as in WWW *all* statements will be non-top-level

Differential Revision: D52513876

fbshipit-source-id: 832161b1b276b56ee3f0379636a5e4cfe849e4a6