Fix references to non-existent `object` type
Summary:
The `object` type does not exist and HHVM interpreted it as any object just by
an accident. Convert them either to the actual type that is being used, or
`\HH\object` if the API accepts/returns any objects, or it was undetermined
which one is being used.

Worth mention is that XSLTProcessor::importStylesheet() was made to no longer
accept SimpleXMLElement (this feature was unused).

Unblocks standard parameter type enforcement for builtins.

Reviewed By: mdko

Differential Revision: D52545395

fbshipit-source-id: 35ae74853fc3b54bc1f0c675912396f89bc95c28