Implement Counter Wrapper Classes
Summary:
This adds the `NullableStats` counter wrapper classes which wrap the `BaseStats` counter classes.

On counter initialization, NullableStats determines the name of the generated counter type and checks if this name is present in the MonitoringConfigLoader regex set. If the name is present, the class is just a passthrough to the BaseStats counter.

Otherwise, it implements a "dummy" counter class that won't instantiate the counter in Service Data and won't add any values when called. This is implemented slightly differently for each counter type due to the underlying counter class and how initialization occurs. But the end result is the same, if the counter name is not present in the monitoring config regex list, it will not be present locally in service data.

Reviewed By: jalopezsilva

Differential Revision: D50521531

fbshipit-source-id: 051089e7af2027a5260e918364fc9f07eaf9226d