capture stdout+stderr of ClientIdeDaemon
Summary:
I'm struggling how to understand why ClientIdeDaemon shuts down sometimes. The previous diff added a bunch of telemetry around its shutdown.

I had one further thought. What if there's some code library which is printing out the reason for its shutdown to stdout or stderr? Previously,
* ClientIdeDaemon's stdin was the same as ClientLsp's stdin, i.e. the LSP protocol message VSCode sends. This was wrong. I suppose we were saved because ClientIdeDaemon never happened to read from stdin.
   * This diff changes it so ClientIdeDaemon's stdin comes from /dev/null.
* ClientIdeDaemon's stdout was the same as ClientLsp's stdout, i.e. the LSP protocol message we send back to VSCode. This was wrong. I suppose we were saved because ClientIdeDaemon never happened to write to stdout.
   * This diff changes it so ClientIdeDaemon's stdout writes to `hh --lsp-logname`
* ClientIdeDaemon's stderr was the same as ClientLsp's stderr, i.e. the stderr channel we send back to VSCode, which who knows what VSCode does with. Does it silently consume it and drop it? Or does it fail to consume it so that eventually the pipe becomes full? I've never known.
   * This diff changes it so ClientIdeDaemon's stderr writes to `hh --lsp-logname`

It can be a bit weird writing to the logfile. That's because if clientLsp process writes to the logfile, and also ClientIdeDaemon writes to the same logfile, then the two might get interleaved. I figure that's okay since we don't even expect anything being written. This diff is solely a last-ditch measure in case something is. The extra overhead of creating an entirely new file just to capture (presumably) non-existing output without interleaving is totally not worth it.

Reviewed By: viratyosin

Differential Revision: D52571102

fbshipit-source-id: c7db77288933becbd8cb4ff64dd3c8e293049d9d