Skip dynamic check for testing standard suite under implicit pessimisation
Summary: To help us see the wood for the trees with differences introduced by pessimisation, run the standard test suite in implicit pessimisation mode without applying the dynamic pass. That way we won't see extra instances of `hh_show` in the output. We have separate tests in `sound_dynamic` for expected errors during the dynamic pass.

Differential Revision: D52495458

fbshipit-source-id: 7a4748413de77c7bc4f8b658543d57a73f1e3611