fix output of Nothing type
Summary:
Noted in the task, the output of some type `nothing` resulted in silly formatting.

Before this diff, the printer would report `|()` when writing out type `nothing`. I believe this is because nothing is actually an empty TUnion (see hackrs/ty/decl/ty.rs comments for TUnion).

This diff also changes `Tintersection([])` to print as `mixed` for parity.

This diff will continue to print a single element `Tunion` as `(|T)` to match the single element `Tintersection`'s `(&T)`. I could have changed both, but rjbailey noted that this ensures we don't hide the fact we're looking at unions/intersections.

Reviewed By: edwinsmith

Differential Revision: D52491369

fbshipit-source-id: 9ef81b22822d114acefdc64b958164761bd842a2