thrift/rust: easy: unit test serde serialization of an empty union
Summary: Adding a unit test to secure the current behavior of empty union serialization via serde and the state of its compatibility with simplejson. Whilst this might or might not be optimal, any changes to this would be a breaking change of compatibility, so having a unit test should make sure any such changes happen in a controlled way.

Reviewed By: dtolnay

Differential Revision: D52604248

fbshipit-source-id: b05bf3c227369206aa00fdfd8cf4b3b2f1eb0efe