Check parent's sort text if current method doesn't set it
Summary:
The diff checks the parent's sort text if the current method doesn't set it. This is done by adding a new function to the `decl_folded_class.ml` file that retrieves the parent's sort text if the current method doesn't have one. The `decl_inherit.ml` file is also updated to use this new function. The `autocomplete_sort_deep_inheritance.php` and `autocomplete_sort_multiple_interfaces.php` files are added to test the new functionality.

decl_inherit was a bit confusing (thanks dlreeves for walking through that code with me!) so as a follow up, I'll be adding a HIP explaining how inheritance of attributes works.

Reviewed By: viratyosin

Differential Revision: D52521358

fbshipit-source-id: 4b7d801906b845dacc1fb66ecce5ceb57133e9dc