Remove special handling for record-replay from ext_std_file
Summary: Remove the special handling for record-replay from the standard file extension. This special handling was only necessary because of a mismatch in how `FileStreamWrapper` handles calls to `access` (which does not resolve file uris) and `open` (which does) and the fact that the extension changes behavior based on the type of the stream wrapper.

Reviewed By: ricklavoie

Differential Revision: D52610792

fbshipit-source-id: 6f9ad1868d9c1e8f3d543d4bae5cb75f768df520