Properly Model EnumClass\Label for Case Types
Summary:
D48687454 introduced a dedicated data type in the runtime for EnumClass Labels. The case types disjointness model needs to be updated to reflect this change.

This diff introduces a new type constructor for `Typing_case_types.Tag` called `LabelData` in order to properly model enumclass labels. At the same time I added another tag `BuiltInData` to represent any runtime data type that is not currently modeled. This in theory should make the analysis more robust for future changes to the runtime in regards to data types, particularly for modeling the set of tags associated with `mixed` or `nonnull`.

Reviewed By: vassilmladenov

Differential Revision: D52632708

fbshipit-source-id: 8c7f1a22c20718fa2f225385050017130754d50d