Ensure local address is updated after connect()
Summary:
Previously, when an AsyncUDPSocket is initialized with a wildcard IP or port 0 for the local address, calling connect() would not update the local address.

This change always updates the local address after connect so that the local address reflects the correct socket address. This corresponds to the behavior in calling bind() without connect().

Reviewed By: frankfeir

Differential Revision: D52493197

fbshipit-source-id: 3db3e691432e2cc76eddc51ae8dfbd51841985ce