rust parser errors, take 2
Summary:
We're going to migrate functions to be case sensitive first; refactor UsedNames
and Strmap to group semantics by symbol type instead of Yes/NoCase.

This revealed that Attributes were incorrectly case-sensitive (they are type names)
but I'm leaving it that way since we eventually want everything to be case sensitive.

Reviewed By: aorenste

Differential Revision: D52657916

fbshipit-source-id: 14adfec07f3ca9ebe629e84d348fa55ee5dfd26e