Fix/improve/delete some pretty printers
Summary:
Some pp were just plainly wrong since they were printing to stout and ignoring the channel parameter. They were anyway not used.

For pp of member flags, we use the same "technique" as for class element and function flags in Typing_defs_flags, converting to a record and deriving `show` on that record.

Reviewed By: madgen

Differential Revision: D52371031

fbshipit-source-id: 8fdd41e708cfff829bea931e6b0ffd552f358457