fix infinite loop in decl_prefetch when file absent
Summary:
While writing unit-tests for decl_provider_prefetch D52655443 I discovered an infinite-loop bug:
1. It has `class C extends A {}` and the naming-table says that A is found in a.php but the disk has been modified (and we haven't yet updated the naming table) so that A actually isn't there.
2. When we're trying to prefetch the folded needs of C, we discover that we need A
3. We're trying to prefetch A, discover it's not in cache, so we put in the queue "please direct-decl-parse a.php and once you're done then revisit A"
4. We direct-decl-parse a.php, and revisit A by repeating step 3.

In the case where a.php fails to include A, then steps 3+4 are an infinite loop.

This diff fixes the bug by only revisiting A if it was indeed provided by direct-decl-parser.

A different solution would have been to rely on "consistency correctness", i.e. throw an exception as soon as we discover that the naming table is out of date, and have that propagate all the way up, and do a watchman query plus re-indexing. But ClientIdeDaemon isn't yet built on top of watchman (is instead at the mercy of whenever VSCode decides to send us a DidChangeWatchedFile notification) so that option isn't available. Also I kind of wanted to make decl_prefetching be a simple best-effort function, not a place to enforce correctness.

Reviewed By: edwinsmith

Differential Revision: D52662856

fbshipit-source-id: 9d11eef3fce5b9ae36912511b1a880d8339379d8