Make call checking with lambda expression arguments less brittle
Summary:
SDT function calling on generic functions is a common source of incompleteness. Here is an example:
```
function reduce<Tv,Ta>(
  Traversable<Tv> $traversable,
  (function (Ta, Tv): Ta) $accumulator,
  Ta $initial,
): Ta {
  ...
}
// Will pessimise as ~IFoo<bool>
function get_foo_bool(): IFoo<bool> {
  ...
}
function testit(): void {
  $f = (IFoo<bool> $b, int $_) ==> $b;
  myreduce(vec[1, 2, 3], $f, get_foo_bool()); // no error
  myreduce(vec[1, 2, 3], (IFoo<bool> $b, int $_):IFoo<bool> ==> $b, get_foo_bool()); // error!
}
```
When checking calls, we first check non-lambda-expression arguments from left-to-right, and then go back and check lambda-expression arguments from left-to-right. Combined with our "non-like then like" SDT checking produces weird incompleteness as seen above. In particular, if there is a parameter than has a generic parameter type (as with `$initial : Ta` above, we will do the static check first, here trying `~IFoo<bool> <: Ta`, and infer a like type for the generic parameter, which is *usually* not the best we can do.

A heuristic that gets us out of this pickle is the following: do the dynamic (like-type) check *first*, if the type of the argument is already a like type, and the parameter type is not a like type. So for the above example we would do `~IFoo<bool> <: ~Ta`, and thereby infer `IFoo<bool>` for `Ta`. The only downside is that we will infer a like type for the result of the function in some circumstances where a fully static call would have been successful.

Differential Revision: D52626747

fbshipit-source-id: 25540e6cf3006f129954ea3a7af5ea4ed3bc96e4